Refactor exchange order and rebalance tests to use fee_msats and fee_original fields
- Updated test cases in `test_exchange_protocol.py` and `test_exchange_rebalance.py` to replace `fee` with `fee_msats` and `fee_original` for better clarity and accuracy in fee representation.
- Added assertions for `fee_original` in relevant tests to ensure correct fee calculations.
- Adjusted the expected values in tests to align with the new fee structure.
- Enhanced logging and error handling in the accounting logic to address issues with zero values and incorrect fee displays.
- Introduced new balance sheet and profit/loss pipelines in `balance_sheet_pipelines_grok.py` to improve financial reporting accuracy.

+++ src/v4vapp_backend_v2/helpers/crypto_prices.py
@@ -150,10 +150,10 @@ class QuoteResponse(BaseModel):
     QuoteResponse is a model that represents the response of cryptocurrency quotes.
 
     Attributes:
-        hive_usd (float): The price of HIVE in USD. Default is 0.
-        hbd_usd (float): The price of HBD in USD. Default is 0.
-        btc_usd (float): The price of BTC in USD. Default is 0.
-        hive_hbd (float): The price of HIVE in HBD. Default is 0.
+        hive_usd (Decimal): The price of HIVE in USD. Default is 0.
+        hbd_usd (Decimal): The price of HBD in USD. Default is 0.
+        btc_usd (Decimal): The price of BTC in USD. Default is 0.
+        hive_hbd (Decimal): The price of HIVE in HBD. Default is 0.
         raw_response (Dict[str, Any]): The raw response data. Default is an
             empty dictionary.
         fetch_date (datetime): The date and time when the data was fetched. Default is
